+++ nata-server/api/v1/api_testcase.js
@@ -1,13 +1,13 @@
 var TestcaseModel = require('../../models/model_testcase.js');
 var _ = require('lodash');
 
-module.exports.create = function(req, res, next) {
+module.exports.create = function (req, res, next) {
   var testcase = new TestcaseModel();
   testcase.name = req.body.name;
   testcase.device_id = req.body.device_id;
-  testcase.apk_id= req.body.apk_id;
+  testcase.apk_id = req.body.apk_id;
 
-  testcase.save(function(err, testcase) {
+  testcase.save(function (err, testcase) {
     if (err) {
       return next(err);
     } else {
@@ -16,11 +16,11 @@ module.exports.create = function(req, res, next) {
   });
 };
 
-module.exports.remove = function(req, res, next) {
+module.exports.remove = function (req, res, next) {
   var testcase_id = req.params.id;
   TestcaseModel.findOneAndRemove({
     _id: testcase_id
-  }, function(err, testcase) {
+  }, function (err, testcase) {
     if (err) {
       return next(err);
     }
@@ -28,15 +28,15 @@ module.exports.remove = function(req, res, next) {
   });
 };
 
-module.exports.finish = function(req, res, next) {
+module.exports.finish = function (req, res, next) {
   var testcase_id = req.params.id;
-  console.log(testcase_id);
   var actions = req.body.actions;
+  console.log(actions);
 
 
   TestcaseModel.findOne({
     _id: testcase_id
-  }, function(err, testcase) {
+  }, function (err, testcase) {
     if (err) {
       next(err);
     }
@@ -46,12 +46,28 @@ module.exports.finish = function(req, res, next) {
     //   testcase.actions.push(action);
     //});
 
-    testcase.isFinish= true;
+    testcase.isFinish = true;
 
-    testcase.save(function(err) {
+    testcase.save(function (err) {
       if (err) return next(err);
       res.status(200).send("success");
     });
+  });
+};
+
+module.exports.save = function (req, res, next) {
+  var testcase_id = req.params.id;
+  var name = req.body.name;
+  var actions = req.body.actions;
+  console.log(name);
+  console.log(actions);
+
+
+  TestcaseModel.findOneAndUpdate({_id: testcase_id}, {name:name,actions:actions},function (err, testcase) {
+    if (err) {
+      next(err);
+    }
+    res.status(200).send("success");
   });
 };