+++ CHANGELOG.md
@@ -165,3 +165,5 @@
 
 ### (2015/5/6)
 - Optimize record testcase
+- Add testcase edit function
+